requirements.txt
streamlit
streamlit-ace
langchain
langchain-community
langchain-google-genai
ollama
chromadb
pypdf
pdfplumber
google-generativeai
typing
